GenericCharacterTables.jl

About

GenericCharacterTables is a library for working with generic character tables. It is based on CHEVIE and aims to provide all its features originally implemented in Maple. It is written in Julia and depends on Nemo and therefore on AbstractAlgebra.

Loading tables

Before doing anything you need to load a table first. GenericCharacterTables comes with a variety of precomputed tables.

GenericCharacterTables.genchartab — Function

genchartab(x::String)

Return the generic character table with name x.

Omitting x will return the names of all importable character tables.

Examples

julia> g = genchartab("GL2")
Generic character table
  of order q^4 - q^3 - q^2 + q
  with 4 irreducible character types
  with 4 class types
  with parameters (i, j, l, k)

GenericCharacterTables also provides some tables of Green functions. They can be loaded similarly.

GenericCharacterTables.greenfuntab — Function

greenfuntab(x::String)

Return the greenfunction table with name x.

Omitting x will return the names of all importable greenfunctions.

Examples

julia> g = greenfuntab("GL2")
Generic character table
  of order q^4 - q^3 - q^2 + q
  with 2 irreducible character types
  with 2 class types
  without parameters