Calculate with Tables

GenericCharacterTables allows you to compute for example scalar products of whole class or character types. The results are correct for all possible combinations of parameters except possibly for those where the additionally returned exceptions apply. Those consist of multivariate polynomials with coefficients in a rational function field and are fulfilled if the evaluation of this polynomial is an integer.

On Characters

LinearAlgebra.norm — Method

norm(char::GenericCharacter)

Return the norm of the character type char.

Examples

julia> g=genchartab("GL2");

julia> norm(g[1])
1

LinearAlgebra.norm — Method

norm(char::SimpleGenericCharacter{T}) where T <: NfPoly

Return the norm of the character type char.

Examples

julia> g=greenfuntab("GL3");

julia> norm(g[1])
6//(q^3 - 3*q^2 + 3*q - 1)

Oscar.scalar_product — Method

scalar_product(char1::GenericCharacter, char2::GenericCharacter)

Return the scalar product between the character types char1 and char2.

Examples

julia> g=genchartab("GL2");

julia> scalar_product(g[3],g[2])
0
With exceptions:
  l1 + k1 - 2*k2 ∈ (q - 1)ℤ
  l1 - k2 ∈ (q - 1)ℤ
  k1 - k2 ∈ (q - 1)ℤ

Oscar.scalar_product — Method

scalar_product(char1::SimpleGenericCharacter{T}, char2::SimpleGenericCharacter{T}) where T <: NfPoly

Return the scalar product between the character types char1 and char2.

Examples

julia> g=greenfuntab("GL3");

julia> scalar_product(g[1],g[2])
0

On Classes

GenericCharacterTables.ortho2norm — Function

ortho2norm(t::CharTable, class::Int64)

Return the (generic) norm of the class type class.

Examples

julia> g=genchartab("GL2");

julia> ortho2norm(g,2)
1

ortho2norm(t::SimpleCharTable{T}, class::Int64) where T <: NfPoly

Return the (generic) norm of the class type class.

Examples

julia> g=greenfuntab("GL3");

julia> ortho2norm(g,2)
(5*q^2 + 2*q + 3)//(q^5 - 2*q^4 + q^3)

GenericCharacterTables.ortho2scalar — Function

ortho2scalar(t::CharTable, class1::Int64, class2::Int64)

Return the (generic) scalar product between the class types class1 and class2.

Examples

julia> g=genchartab("GL2");

julia> ortho2scalar(g,3,2)
0
With exceptions:
  i1 + j1 - 2*i2 ∈ (q - 1)ℤ
  i1 - i2 ∈ (q - 1)ℤ
  j1 - i2 ∈ (q - 1)ℤ

ortho2scalar(t::SimpleCharTable{T}, class1::Int64, class2::Int64) where T <: NfPoly

Return the (generic) scalar product between the class types class1 and class2.

Examples

julia> g=greenfuntab("GL3");

julia> ortho2scalar(g,2,2)
(5*q^2 + 2*q + 3)//(q^5 - 2*q^4 + q^3)

GenericCharacterTables.classmult — Function

classmult(t::CharTable, class1::Int64, class2::Int64, class3::Int64)

Return the (generic) class multiplication constant of the class types class1, class2 and class3 of the table t.

Examples

julia> g=genchartab("SL2.0");

julia> classmult(g,2,2,4)
q + 1
With exceptions:
  a3 ∈ (q + 1)ℤ

classmult(t::SimpleCharTable{T}, class1::Int64, class2::Int64, class3::Int64) where T <: NfPoly

Return the (generic) class multiplication constant of the class types class1, class2 and class3 of the table t.

Examples

julia> g=greenfuntab("GL3");

julia> classmult(g,1,2,3)
(q + 3)//(q^5 - 2*q^4 + q^3)