Character types

A character type is a family of characters which are indexed by a set of parameters, together with ranges of admissible values for each parameters, and a set of excluded parameter values. The characters in a character type share many properties, e.g. they all have the same degree.

Since "types" already have a very specific meaning in Julia (and other programming languages), we instead sometimes refer to character types as "generic characters". In particular the Julia types we use to represent character types are called AbstractGenericCharacter, GenericCharacter and SimpleGenericCharacter.

Properties

GenericCharacterTables.number_of_character_types — Function

number_of_character_types(t::Table)

Return the number of character types of table t. This can also be obtained via length(t).

Examples

julia> g=generic_character_table("GL2");

julia> number_of_character_types(g)
4

GenericCharacterTables.number_of_characters — Method

number_of_characters(char::GenericCharacter)

Return the number of characters in the generic character char.

Examples

julia> g=generic_character_table("GL2");

julia> number_of_characters(g[1])
q - 1

AbstractAlgebra.degree — Method

degree(char::AbstractGenericCharacter)

Return the character degree of the characters in char.

Examples

julia> g=generic_character_table("GL2");

julia> degree(g[3])
q + 1

GenericCharacterTables.parameters — Method

parameters(char::AbstractGenericCharacter)

Return the parameters of the character type char. This includes the parameter names, ranges and exceptions.

Examples

julia> g=generic_character_table("GL2");

julia> parameters(g[3])
k ∈ {1,…, q - 1}, l ∈ {1,…, q - 1} except -l + k ∈ (q - 1)ℤ

GenericCharacterTables.info — Method

info(char::AbstractGenericCharacter)

Return the infolists of the character type char.

Examples

julia> g=generic_character_table("GL2");

julia> info(g[2])
2-element Vector{Any}:
 Any[1, 1]
 Any["A_1", [1, 1]]

Iteration

Character types behave in some ways similar to Julia Vector objects in that one can index into them to extract values. For a character type ct,

length(ct) returns the number of values in the character type (which is equal to the number of class types s in the table),

ct[i] returns the $i$th value of the character type,

ct[i:j] returns the values $i$ to $j$,

ct[:] returns all values.

The $i$th character type can be obtained from a generic character table g in multiple ways: g[i], g[i,:] or character_type(g,i). The following example demonstrates this.

julia> g=generic_character_table("GL2");

julia> character_type(g,3)
Generic character of GL2
  with parameters
    k ∈ {1,…, q - 1}, l ∈ {1,…, q - 1} except -l + k ∈ (q - 1)ℤ
  of degree q + 1
  with values
    1: (q + 1)*E(q - 1)^(i*l + i*k)
    2: E(q - 1)^(i*l + i*k)
    3: E(q - 1)^(i*l + j*k) + E(q - 1)^(i*k + j*l)
    4: 0

julia> g[3,:]
Generic character of GL2
  with parameters
    k ∈ {1,…, q - 1}, l ∈ {1,…, q - 1} except -l + k ∈ (q - 1)ℤ
  of degree q + 1
  with values
    1: (q + 1)*E(q - 1)^(i*l + i*k)
    2: E(q - 1)^(i*l + i*k)
    3: E(q - 1)^(i*l + j*k) + E(q - 1)^(i*k + j*l)
    4: 0

julia> ct = g[3]
Generic character of GL2
  with parameters
    k ∈ {1,…, q - 1}, l ∈ {1,…, q - 1} except -l + k ∈ (q - 1)ℤ
  of degree q + 1
  with values
    1: (q + 1)*E(q - 1)^(i*l + i*k)
    2: E(q - 1)^(i*l + i*k)
    3: E(q - 1)^(i*l + j*k) + E(q - 1)^(i*k + j*l)
    4: 0

julia> ct[1]
(q + 1)*E(q - 1)^(i*l + i*k)

julia> ct[1:3]
3-element Vector{GenericCharacterTables.GenericCyclo}:
 (q + 1)*E(q - 1)^(i*l + i*k)
 E(q - 1)^(i*l + i*k)
 E(q - 1)^(i*l + j*k) + E(q - 1)^(i*k + j*l)

julia> ct[2:end]
3-element Vector{GenericCharacterTables.GenericCyclo}:
 E(q - 1)^(i*l + i*k)
 E(q - 1)^(i*l + j*k) + E(q - 1)^(i*k + j*l)
 0

julia> collect(ct)
4-element Vector{GenericCharacterTables.GenericCyclo}:
 (q + 1)*E(q - 1)^(i*l + i*k)
 E(q - 1)^(i*l + i*k)
 E(q - 1)^(i*l + j*k) + E(q - 1)^(i*k + j*l)
 0

julia> ct[:]
4-element Vector{GenericCharacterTables.GenericCyclo}:
 (q + 1)*E(q - 1)^(i*l + i*k)
 E(q - 1)^(i*l + i*k)
 E(q - 1)^(i*l + j*k) + E(q - 1)^(i*k + j*l)
 0

Parameter names

When two characters types are for example added together or multiplied with each other, it is necessary to differentiate the parameters of the two character types. Otherwise the result will only represent a fraction of all possible combinations of characters. Namely those where the parameters of the character types with identical name have the same value. Therefore the parameter k will be called kti whenever it represents the parameter k of the $i$-th character type of a tensor product. Similarly it will be called kli in a linear combination and ki in a scalar product.

Note that for linear combinations this has the possibly unexpected implication that $2\chi$ represents the character type containing the double of each character contained in $\chi$ while $\chi+\chi$ represents the character type containing every possible sum of two characters in $\chi$.

Constructing new character types

The immediate way to obtain a character type object is to get it from a table T via indexing, i.e., as T[i] for some index $i$.

In addition, there are a few ways to construct new character types.

Hecke.tensor_product — Method

tensor_product(char1::GenericCharacter, char2::GenericCharacter)

Return the tensor product of the character types char1 and char2. This can also be obtained via char1 * char2.

Examples

julia> g = generic_character_table("GL2");

julia> tensor_product(g[1], g[2])
Generic character of GL2
  with parameters
    kt1 ∈ {1,…, q - 1}, kt2 ∈ {1,…, q - 1}
  of degree q
  with values
    1: q*E(q - 1)^(2*i*kt1 + 2*i*kt2)
    2: 0
    3: E(q - 1)^(i*kt1 + i*kt2 + j*kt1 + j*kt2)
    4: -E(q - 1)^(i*kt1 + i*kt2)

Hecke.tensor_product — Method

tensor_product(char1::SimpleGenericCharacter{T}, char2::SimpleGenericCharacter{T}) where T<:PolyRingElem

Return the tensor product of the character types char1 and char2. This can also be obtained via char1 * char2.

Examples

julia> g = green_function_table("GL3");

julia> tensor_product(g[1],g[2])
Generic character of GL3
  of degree -q^6 - 2*q^5 - 2*q^4 + 2*q^2 + 2*q + 1
  with values
    1: -q^6 - 2*q^5 - 2*q^4 + 2*q^2 + 2*q + 1
    2: 2*q + 1
    3: 1

GenericCharacterTables.linear_combination — Function

linear_combination(coeffs::Vector{<:RingElement}, chars::Vector{<:GenericCharacter})

Return the linear combination of the character types chars with coefficients coeffs.

Examples

julia> g=generic_character_table("GL2");

julia> linear_combination([5,1],[g[1],g[2]])
Generic character of GL2
  with parameters
    kl1 ∈ {1,…, q - 1}, kl2 ∈ {1,…, q - 1}
  of degree q + 5
  with values
    1: 5*E(q - 1)^(2*i*kl1) + q*E(q - 1)^(2*i*kl2)
    2: 5*E(q - 1)^(2*i*kl1)
    3: E(q - 1)^(i*kl2 + j*kl2) + 5*E(q - 1)^(i*kl1 + j*kl1)
    4: -E(q - 1)^(i*kl2) + 5*E(q - 1)^(i*kl1)

linear_combination(coeffs::Vector{<:RingElement}, chars::Vector{SimpleGenericCharacter{T}}) where T <: NfPoly

Return the linear combination of the character types chars with coefficients coeffs.

Examples

julia> g=green_function_table("GL3");

julia> linear_combination([5,1],[g[1],g[2]])
Generic character of GL3
  of degree 4*q^3 + 10*q^2 + 10*q + 6
  with values
    1: 4*q^3 + 10*q^2 + 10*q + 6
    2: 10*q + 6
    3: 6

GenericCharacterTables.omega — Function

omega(char::GenericCharacter)

Return the (generic) central character of the character type char.

Examples

julia> g=generic_character_table("GL2");

julia> omega(g[1])
Generic character of GL2
  with parameters
    k ∈ {1,…, q - 1}
  of degree 1
  with values
    1: E(q - 1)^(2*i*k)
    2: (q^2 - 1)*E(q - 1)^(2*i*k)
    3: (q^2 + q)*E(q - 1)^(i*k + j*k)
    4: (q^2 - q)*E(q - 1)^(i*k)

omega(char::SimpleGenericCharacter{T}) where T <: NfPoly

Return the (generic) central character of the character type char.

Examples

julia> g=green_function_table("GL3");

julia> omega(g[1])
Generic character of GL3
  of degree 1
  with values
    1: 1
    2: 2*q^2 - q - 1
    3: q^3 - 2*q^2 + q

Norms and scalar products

GenericCharacterTables allows you to compute norms and scalar products of character types. The results are correct for all possible combinations of parameters except possibly for those where the additionally returned exceptions apply. Those consist of multivariate polynomials with coefficients in a rational function field and are satisfied if the evaluation of this polynomial is an integer.

LinearAlgebra.norm — Method

norm(char::GenericCharacter)

Return the norm of the character type char.

Examples

julia> g=generic_character_table("GL2");

julia> norm(g[1])
1

LinearAlgebra.norm — Method

norm(char::SimpleGenericCharacter{T}) where T <: NfPoly

Return the norm of the character type char.

Examples

julia> g=green_function_table("GL3");

julia> norm(g[1])
6//(q^3 - 3*q^2 + 3*q - 1)

Oscar.scalar_product — Method

scalar_product(char1::GenericCharacter, char2::GenericCharacter)

Return the scalar product between the character types char1 and char2.

Examples

julia> g=generic_character_table("GL2");

julia> scalar_product(g[3],g[2])
0
With exceptions:
  l1 + k1 - 2*k2 ∈ (q - 1)ℤ
  l1 - k2 ∈ (q - 1)ℤ
  k1 - k2 ∈ (q - 1)ℤ

Oscar.scalar_product — Method

scalar_product(char1::SimpleGenericCharacter{T}, char2::SimpleGenericCharacter{T}) where T <: NfPoly

Return the scalar product between the character types char1 and char2.

Examples

julia> g=green_function_table("GL3");

julia> scalar_product(g[1],g[2])
0