Conjugacy class types

A class type is a family of conjugacy classes which are indexed by a set of parameters, together with ranges of admissible values for each parameters, and a set of excluded parameter values.

Properties

GenericCharacterTables.number_of_conjugacy_class_types — Function

number_of_conjugacy_class_types(t::Table)

Return the number of conjugacy class types of table t.

Examples

julia> g=generic_character_table("GL2");

julia> number_of_conjugacy_class_types(g)
4

Oscar.number_of_conjugacy_classes — Method

number_of_conjugacy_classes(t::Table)

Return the number of conjugacy classes of table t.

Examples

julia> g=generic_character_table("GL2");

julia> number_of_conjugacy_classes(g)
q^2 - 1

Oscar.number_of_conjugacy_classes — Method

number_of_conjugacy_classes(t::CharTable, class::Int64)

Return the number of conjugacy classes in the class type class of the table t.

Examples

julia> g=generic_character_table("GL2");

julia> number_of_conjugacy_classes(g, 1)
q - 1

Oscar.number_of_conjugacy_classes — Method

number_of_conjugacy_classes(t::SimpleCharTable, class::Int64)

Return the number of conjugacy classes in the class type class of the table t.

Examples

julia> g=green_function_table("GL2");

julia> number_of_conjugacy_classes(g, 1)
1

GenericCharacterTables.centralizer_order — Function

centralizer_order(t::Table, class::Int64)

Return the order of the centralizer of the class type class of the table t.

Examples

julia> g=generic_character_table("GL2");

julia> centralizer_order(g, 1)
q^4 - q^3 - q^2 + q

GenericCharacterTables.parameters — Method

parameters(t::CharTable, class::Int64)

Return the parameters of the class type class of the table t. This includes the parameter names, ranges and exceptions.

Examples

julia> g=generic_character_table("GL2");

julia> parameters(g, 3)
i ∈ {1,…, q - 1}, j ∈ {1,…, q - 1} except i - j ∈ (q - 1)ℤ

GenericCharacterTables.info — Method

info(t::Table, class::Int64)

Return the infolists of the class type class of the table t.

Examples

julia> g=generic_character_table("GL2");

julia> info(g, 1)
2-element Vector{Any}:
 Any[1, 0]
 Any["A_1", [1, 1]]

Norms and scalar products

GenericCharacterTables allows you to compute norms and scalar products of class types. The results are correct for all possible combinations of parameters except possibly for those where the additionally returned exceptions apply. Those consist of multivariate polynomials with coefficients in a rational function field and are satisfied if the evaluation of this polynomial is an integer.

LinearAlgebra.norm — Method

norm(t::CharTable, class::Int64)

Return the (generic) norm of the class type class.

Examples

julia> g=generic_character_table("GL2");

julia> norm(g,2)
1

LinearAlgebra.norm — Method

norm(t::SimpleCharTable{T}, class::Int64) where T <: NfPoly

Return the (generic) norm of the class type class.

Examples

julia> g=green_function_table("GL3");

julia> norm(g,2)
(5*q^2 + 2*q + 3)//(q^5 - 2*q^4 + q^3)

Oscar.scalar_product — Method

scalar_product(t::CharTable, class1::Int64, class2::Int64)

Return the (generic) scalar product between the class types class1 and class2.

Examples

julia> g=generic_character_table("GL2");

julia> scalar_product(g,3,2)
0
With exceptions:
  i1 + j1 - 2*i2 ∈ (q - 1)ℤ
  i1 - i2 ∈ (q - 1)ℤ
  j1 - i2 ∈ (q - 1)ℤ

Oscar.scalar_product — Method

scalar_product(t::SimpleCharTable{T}, class1::Int64, class2::Int64) where T <: NfPoly

Return the (generic) scalar product between the class types class1 and class2.

Examples

julia> g=green_function_table("GL3");

julia> scalar_product(g,2,2)
(5*q^2 + 2*q + 3)//(q^5 - 2*q^4 + q^3)

Oscar.class_multiplication_coefficient — Method

class_multiplication_coefficient(t::CharTable, class1::Int64, class2::Int64, class3::Int64)

Return the (generic) class multiplication coefficient of the class types class1, class2 and class3 of the table t.

Examples

julia> g=generic_character_table("SL2.0");

julia> class_multiplication_coefficient(g,2,2,4)
q + 1
With exceptions:
  a3 ∈ (q + 1)ℤ

Oscar.class_multiplication_coefficient — Method

class_multiplication_coefficient(t::SimpleCharTable{T}, class1::Int64, class2::Int64, class3::Int64) where T <: NfPoly

Return the (generic) class multiplication coefficient of the class types class1, class2 and class3 of the table t.

Examples

julia> g=green_function_table("GL3");

julia> class_multiplication_coefficient(g,1,2,3)
(q + 3)//(q^5 - 2*q^4 + q^3)