Cyclotomic factorization

Sometimes it can be interesting to see which cyclotomic polynomial divides a given polynomial representing a group or centralizer order.

GenericCharacterTables.factor_cyclotomic — Function

factor_cyclotomic(p::PolyRingElem)

Return a factorization of p into cyclotomic polynomials. The main difference to factor is the way the result gets printed.

Examples

julia> g=green_function_table("GL6");

julia> factor_cyclotomic(order(g))
q^15*Φ_5*Φ_4*Φ_6*Φ_2^3*Φ_3^2*Φ_1^6

factor_cyclotomic(p::Generic.UnivPoly)

Return a factorization of p into cyclotomic polynomials. The main difference to factor is the way the result gets printed.

Examples

julia> g=generic_character_table("GL3");

julia> factor_cyclotomic(order(g))
q^3*Φ_2*Φ_3*Φ_1^3