Get Information

About Tables

GenericCharacterTables.info — Function

info(t::Table)

Return the metadata of t in LaTeX format. This usually includes the time the table was first computed.

GenericCharacterTables.nrparams — Function

nrparams(t::CharTable)

Return the number of class and character parameters of the table t.

Examples

julia> g=genchartab("GL2");

julia> nrparams(g)
4

GenericCharacterTables.params — Function

params(t::CharTable)

Return all parameters the table t depends on.

Examples

julia> g=genchartab("GL2");

julia> params(g)
(q, (i, j, l, k))

params(chi::AbstractGenericCharacter)

Return the parameters of the character type chi. This includes the parameter names, ranges and exceptions.

Examples

julia> g=genchartab("GL2");

julia> params(g[3])
k ∈ {1,…, q - 1}, l ∈ {1,…, q - 1} except -l + k ∈ (q - 1)ℤ

AbstractAlgebra.order — Method

order(t::Table)

Return the order of the table t.

Examples

julia> g=genchartab("GL2");

julia> order(g)
q^4 - q^3 - q^2 + q

About Characters

GenericCharacterTables.nrchars — Function

nrchars(char::GenericCharacter)

Return the number of characters in the generic character char.

Examples

julia> g=genchartab("GL2");

julia> nrchars(g[1])
q - 1

GenericCharacterTables.nrirrchars — Function

nrirrchars(t::Table)

Return the number of irreducible characters of table t.

Examples

julia> g=genchartab("GL2");

julia> nrirrchars(g)
q^2 - 1

AbstractAlgebra.degree — Method

degree(char::AbstractGenericCharacter)

Return the character degree of char.

Examples

julia> g=genchartab("GL2");

julia> degree(g[3])
q + 1

GenericCharacterTables.params — Method

params(chi::AbstractGenericCharacter)

Return the parameters of the character type chi. This includes the parameter names, ranges and exceptions.

Examples

julia> g=genchartab("GL2");

julia> params(g[3])
k ∈ {1,…, q - 1}, l ∈ {1,…, q - 1} except -l + k ∈ (q - 1)ℤ

GenericCharacterTables.printinfochar — Function

printinfochar([io::IO], t::Table, char::Union{Int64, Nothing}=nothing)

Print to io (or to the default output stream stdout if io is not given) the infolists of the character type char of the table t.

Leaving char unspecified will print the infolists of all character types.

Examples

julia> g=genchartab("GL2");

julia> printinfochar(g)
1	Any[Any[1, 0], Any["A_1", [2]]]
2	Any[Any[1, 1], Any["A_1", [1, 1]]]
3	Any[Any[2, 0], Any["A_0", [1]]]
4	Any[Any[3, 0], Any["A_0", [1]]]

About Classes

GenericCharacterTables.classtypes — Function

classtypes(t::Table)

Return the number of conjugacy class types of table t.

Examples

julia> g=genchartab("GL2");

julia> classtypes(g)
4

GenericCharacterTables.nrclasses — Function

nrclasses(t::Table)

Return the number of conjugacy classes of table t.

Examples

julia> g=genchartab("GL2");

julia> nrclasses(g)
q^2 - 1

nrclasses(t::Table, class::Int64)

Return the number of conjugacy classes in the class type class of the table t.

Examples

julia> g=genchartab("GL2");

julia> nrclasses(g, 1)
q - 1

GenericCharacterTables.centord — Function

centord(t::Table, class::Int64)

Return the order of the centralizer of the class type class of the table t.

Examples

julia> g=genchartab("GL2");

julia> centord(g, 1)
q^4 - q^3 - q^2 + q

GenericCharacterTables.printclassparam — Function

printclassparam([io::IO], t::CharTable, class::Union{Int64, Nothing}=nothing)

Print to io (or to the default output stream stdout if io is not given) the parameters of the class type class of the table t.

This includes the parameter names, ranges and exceptions. Leaving class unspecified will print the parameters of all character types.

Examples

julia> g=genchartab("GL2");

julia> printclassparam(g)
1	i ∈ {1,…, q - 1}
2	i ∈ {1,…, q - 1}
3	i ∈ {1,…, q - 1}, j ∈ {1,…, q - 1} except i - j ∈ (q - 1)ℤ
4	i ∈ {1,…, q^2 - 1} except i ∈ (q + 1)ℤ

GenericCharacterTables.printinfoclass — Function

printinfoclass([io::IO], t::Table, class::Union{Int64, Nothing}=nothing)

Print to io (or to the default output stream stdout if io is not given) the infolists of the class type class of the table t.

Leaving class unspecified will print the infolists of all class types.

Examples

julia> g=genchartab("GL2");

julia> printinfoclass(g)
1	Any[Any[1, 0], Any["A_1", [1, 1]]]
2	Any[Any[1, 1], Any["A_1", [2]]]
3	Any[Any[2, 0], Any["A_0", [1]]]
4	Any[Any[3, 0], Any["A_0", [1]]]