Tables

This section describes how to instantiate generic character tables and tables of Green functions, as well as various functions for accessing properties of these tables.

Loading tables

Before doing anything you need to load a table first. GenericCharacterTables comes with a variety of precomputed tables.

GenericCharacterTables.genchartab — Function

genchartab(x::String)

Return the generic character table with name x.

Omitting x will return the names of all importable character tables.

Examples

julia> g = genchartab("GL2")
Generic character table GL2
  of order q^4 - q^3 - q^2 + q
  with 4 irreducible character types
  with 4 class types
  with parameters (i, j, l, k)

GenericCharacterTables also provides some tables of Green functions. They can be loaded similarly.

GenericCharacterTables.greenfuntab — Function

greenfuntab(x::String)

Return the greenfunction table with name x.

Omitting x will return the names of all importable greenfunctions.

Examples

julia> g = greenfuntab("GL2")
Generic character table GL2
  of order q^4 - q^3 - q^2 + q
  with 2 irreducible character types
  with 2 class types
  without parameters

Properties

GenericCharacterTables.info — Method

info(t::Table)

Return the metadata of t in LaTeX format. This usually includes the time the table was first computed.

GenericCharacterTables.number_of_characters — Method

number_of_characters(t::Table)

Return the number of irreducible characters of table t.

Examples

julia> g=genchartab("GL2");

julia> number_of_characters(g)
q^2 - 1

GenericCharacterTables.nrparams — Function

nrparams(t::CharTable)

Return the number of class and character parameters of the table t.

Examples

julia> g=genchartab("GL2");

julia> nrparams(g)
4

GenericCharacterTables.params — Function

params(t::CharTable, class::Int64)

Return the parameters of the class type class of the table t. This includes the parameter names, ranges and exceptions.

Examples

julia> g=genchartab("GL2");

julia> params(g, 3)
i ∈ {1,…, q - 1}, j ∈ {1,…, q - 1} except i - j ∈ (q - 1)ℤ

params(t::CharTable)

Return all parameters the table t depends on.

Examples

julia> g=genchartab("GL2");

julia> params(g)
(q, (i, j, l, k))

params(char::AbstractGenericCharacter)

Return the parameters of the character type char. This includes the parameter names, ranges and exceptions.

Examples

julia> g=genchartab("GL2");

julia> params(g[3])
k ∈ {1,…, q - 1}, l ∈ {1,…, q - 1} except -l + k ∈ (q - 1)ℤ

AbstractAlgebra.order — Method

order(t::Table)

Return the order of the table t.

Examples

julia> g=genchartab("GL2");

julia> order(g)
q^4 - q^3 - q^2 + q

Iteration

Tables implement Julia's iteration interface to iterate over the irreducible character types stored in the table. For a table T,

length(T) returns the number of character types in the table,

T[i] returns the $i$th character type.

For example we can use this to compute the order of the underlying group type. (Of course this can also be checked via order, which retrieves a precomputed value.)

julia> g=genchartab("GL2");

julia> sum(number_of_characters(c)*degree(c)^2 for c in g)
q^4 - q^3 - q^2 + q

julia> order(g)
q^4 - q^3 - q^2 + q