Unexported Functions

These are all functions for internal use only:

GenericCharacterTables.CharTable — Method

(t::CharTable)(c::GenericCharacter)

Return c as a generic character of t. This will only work if t is a version of the parent table of c with a more restricted congruence.

AbstractAlgebra.Generic.normal_form — Method

normal_form(f::AbstractAlgebra.Generic.UnivPoly{ZZRingElem, AbstractAlgebra.Generic.MPoly{ZZRingElem}}, m::Int64)

Return a normal form of f modulo m, such that normal_form(f,m) is equal to normal_form(g,m) if and only if f and g are congruent modulo m.

Examples

julia> R=universal_polynomial_ring(ZZ);

julia> x=gen(R, :x);

julia> normal_form(4*x^2,6)
x^2 + 3*x

julia> normal_form(4*x^2-(x^2+3*x),6)
0

julia> normal_form(4*x^9+x^7,12)
x^3 + 4*x

julia> normal_form(4*x^9+x^7-(x^3+4*x),12)
0

Base.iszero — Method

iszero(x::GenericCycloFrac; ignore_exceptions::Bool=false)

Return if x is zero. If ignore_exceptions is true then the exceptions of x will not be considered.

Base.show — Method

show(io::IO, c::AbstractGenericCharacter)

Display a summary of the generic character c.

Examples

julia> g=genchartab("GL2");

julia> g[3]
Generic character of GL2
  with parameters
    k ∈ {1,…, q - 1}, l ∈ {1,…, q - 1} except -l + k ∈ (q - 1)ℤ
  of degree q + 1
  with values
    (q + 1)*exp(2π𝑖((i*l + i*k)//(q - 1)))
    exp(2π𝑖((i*l + i*k)//(q - 1)))
    exp(2π𝑖((i*l + j*k)//(q - 1))) + exp(2π𝑖((i*k + j*l)//(q - 1)))
    0

julia> [g[3]]
1-element Vector{GenericCharacterTables.GenericCharacter}:
 Generic character of GL2

Base.show — Method

show(io::IO, t::Table)

Display a summary of the generic character table t.

Examples

julia> g=genchartab("GL2")
Generic character table GL2
  of order q^4 - q^3 - q^2 + q
  with 4 irreducible character types
  with 4 class types
  with parameters (i, j, l, k)

julia> [g]
1-element Vector{GenericCharacterTables.CharTable}:
 Generic character table GL2

GenericCharacterTables.shift_char_parameters — Method

shift_char_parameters(t::CharTable, a::Union{Parameters,GenericCyclo,GenericCycloFrac}, steps::Int64)

Replace all character parameters of t in a by their counterparts suffixed with steps.

This is done by shifting them steps*nrparams(t) steps further in t.argumentring.

GenericCharacterTables.shift_class_parameters — Method

shift_class_parameters(t::CharTable, a::Union{Parameters,GenericCyclo,GenericCycloFrac}, steps::Int64)

Replace all class parameters of t in a by their counterparts suffixed with steps.

This is done by shifting them steps*nrparams(t) steps further in t.argumentring.

GenericCharacterTables.shrink — Method

shrink(a::GenericCycloFrac{<:NfPoly})

Remove exceptions from a that follow from the others. And try to simplify the representation of a.