Get Information

About Tables

GenericCharacterTables.status — Function

status(t::Table)

Print the order of the group associated to t and the number of class and character types of the table t.

Examples

julia> g=genchartab("GL2");

julia> status(g)
Order:                     q^4 - q^3 - q^2 + q
Number of character types: 4
Number of class types:     4

GenericCharacterTables.printval — Function

printval(t::Table; char::Union{Int64, Nothing}=nothing, class::Union{Int64, Nothing}=nothing)

Print the values of the char type char and the class type class of the table t.

Leaving class unspecified will print the values of all class types at char. Leaving both unspecified will print all values of t

Examples

julia> g=genchartab("GL2");

julia> printval(g,char=1)
Value of character type 1 on class type
  1: exp(2π𝑖(2//(q - 1)*i*k))
  2: exp(2π𝑖(2//(q - 1)*i*k))
  3: exp(2π𝑖(1//(q - 1)*i*k + 1//(q - 1)*j*k))
  4: exp(2π𝑖(1//(q - 1)*i*k))

julia> printval(g,char=4,class=2)
Value of character type 4 on class type
  2: -1*exp(2π𝑖(1//(q - 1)*i*k))

GenericCharacterTables.printinfotab — Function

printinfotab(t::Table)

Print metadata of t in the latex format.

This usually includes the time the table was first computed.

GenericCharacterTables.nrparams — Function

nrparams(t::CharTable)

Return the number of class and character parameters of the table t.

Examples

julia> g=genchartab("GL2");

julia> nrparams(g)
4

GenericCharacterTables.params — Function

params(t::CharTable)

Return all parameters the table t depends on.

Examples

julia> g=genchartab("GL2");

julia> params(g)
(q, (i, j, l, k))

AbstractAlgebra.order — Method

order(t::Table)

Return the order of the table t.

Examples

julia> g=genchartab("GL2");

julia> order(g)
q^4 - q^3 - q^2 + q

About Characters

GenericCharacterTables.chartypes — Function

chartypes(t::Table)

Return the number of character types of table t.

Examples

julia> g=genchartab("GL2");

julia> chartypes(g)
4

GenericCharacterTables.irrchartypes — Function

irrchartypes(t::Table)

Return the number of irreducible character types of table t.

For example this excludes character types created with tensor! or lincomb!.

GenericCharacterTables.nrchars — Function

nrchars(t::Table{T}, char::Int64) where T <: NfPoly

Return the number of characters in the character type char of the table t.

In the case of a SimpleCharTable this is always one.

Examples

julia> g=genchartab("GL2");

julia> nrchars(g, 1)
q - 1

GenericCharacterTables.nrirrchars — Function

nrirrchars(t::Table)

Return the number of irreducible characters of table t.

For example this excludes characters created with tensor! or lincomb!.

Examples

julia> g=genchartab("GL2");

julia> nrirrchars(g)
q^2 - 1

GenericCharacterTables.chardeg — Function

chardeg(t::Table, char::Int64)

Return the character degree of the character type char of the table t.

Examples

julia> g=genchartab("GL2");

julia> chardeg(g, 3)
q + 1

GenericCharacterTables.printcharparam — Function

printcharparam(t::CharTable, char::Union{Int64, Nothing}=nothing)

Print the parameters of the character type char of the table t.

This includes the parameter names, ranges and exceptions. Leaving char unspecified will print the parameters of all character types.

Examples

julia> g=genchartab("GL2");

julia> printcharparam(g)
1	k ∈ {1,…, q - 1}
2	k ∈ {1,…, q - 1}
3	k ∈ {1,…, q - 1}, l ∈ {1,…, q - 1} except (l - k)//(q - 1) ∈ ℤ
4	k ∈ {1,…, q^2 - 1} except (k)//(q + 1) ∈ ℤ

GenericCharacterTables.printinfochar — Function

printinfochar(t::Table, char::Union{Int64, Nothing}=nothing)

Print the infolists of the character type char of the table t.

Leaving char unspecified will print the infolists of all character types.

Examples

julia> g=genchartab("GL2");

julia> printinfochar(g)
1	Any["", [1, 0], Any["A_1", [2]]]
2	Any["", [1, 1], Any["A_1", [1, 1]]]
3	Any["", [2, 0], Any["A_0", [1]]]
4	Any["", [3, 0], Any["A_0", [1]]]

About Classes

GenericCharacterTables.classtypes — Function

classtypes(t::Table)

Return the number of conjugacy class types of table t.

Examples

julia> g=genchartab("GL2");

julia> classtypes(g)
4

GenericCharacterTables.nrclasses — Function

nrclasses(t::Table)

Return the number of conjugacy classes of table t.

Examples

julia> g=genchartab("GL2");

julia> nrclasses(g)
q^2 - 1

nrclasses(t::Table{T}, class::Int64) where T <: NfPoly

Return the number of conjugacy classes in the class type class of the table t.

Examples

julia> g=genchartab("GL2");

julia> nrclasses(g, 1)
q - 1

GenericCharacterTables.centord — Function

centord(t::Table, class::Int64)

Return the order of the centralizer of the class type class of the table t.

Examples

julia> g=genchartab("GL2");

julia> centord(g, 1)
q^4 - q^3 - q^2 + q

GenericCharacterTables.printclassparam — Function

printclassparam(t::CharTable, class::Union{Int64, Nothing}=nothing)

Print the parameters of the class type class of the table t.

This includes the parameter names, ranges and exceptions. Leaving class unspecified will print the parameters of all character types.

Examples

julia> g=genchartab("GL2");

julia> printclassparam(g)
1	i ∈ {1,…, q - 1}
2	i ∈ {1,…, q - 1}
3	i ∈ {1,…, q - 1}, j ∈ {1,…, q - 1} except (i - j)//(q - 1) ∈ ℤ
4	i ∈ {1,…, q^2 - 1} except (i)//(q + 1) ∈ ℤ

GenericCharacterTables.printinfoclass — Function

printinfoclass(t::Table, class::Union{Int64, Nothing}=nothing)

Print the infolists of the class type class of the table t.

Leaving class unspecified will print the infolists of all class types.

Examples

julia> g=genchartab("GL2");

julia> printinfoclass(g)
1	Any["", [1, 0], Any["A_1", [1, 1]]]
2	Any["", [1, 1], Any["A_1", [2]]]
3	Any["", [2, 0], Any["A_0", [1]]]
4	Any["", [3, 0], Any["A_0", [1]]]