Tables

This section describes how to instantiate generic character tables and tables of Green functions, as well as various functions for accessing properties of these tables.

Loading tables

Before doing anything you need to load a table first. GenericCharacterTables comes with a variety of precomputed tables.

GenericCharacterTables.generic_character_table — Function

generic_character_table(x::String)

Return the generic character table with name x.

Omitting x will return the names of all importable character tables.

Examples

julia> g = generic_character_table("GL2")
Generic character table GL2
  of order q^4 - q^3 - q^2 + q
  with 4 irreducible character types
  with 4 class types
  with parameters (i, j, l, k)

GenericCharacterTables also provides some tables of Green functions. They can be loaded similarly.

GenericCharacterTables.green_function_table — Function

green_function_table(x::String)

Return the greenfunction table with name x.

Omitting x will return the names of all importable greenfunctions.

Examples

julia> g = green_function_table("GL2")
Generic character table GL2
  of order q^4 - q^3 - q^2 + q
  with 2 irreducible character types
  with 2 class types
  without parameters

Properties

GenericCharacterTables.info — Method

info(t::Table)

Return the metadata of t in LaTeX format. This usually includes the time the table was first computed.

Examples

julia> g=generic_character_table("GL2");

julia> info(g)
  The generic character table of \mathrm{GL}_2(q).

    •  The table was first computed in Jor07 (@cite), Sch07 (@cite).

    •  See also: Ste51 (@cite).

GenericCharacterTables.number_of_characters — Method

number_of_characters(t::Table)

Return the number of irreducible characters of table t.

Examples

julia> g=generic_character_table("GL2");

julia> number_of_characters(g)
q^2 - 1

GenericCharacterTables.number_of_parameters — Function

number_of_parameters(t::CharTable)

Return the number of class and character parameters of the table t.

Examples

julia> g=generic_character_table("GL2");

julia> number_of_parameters(g)
4

number_of_parameters(t::SimpleCharTable)

Return the number of class and character parameters of the table t.

Examples

julia> g=green_function_table("GL2");

julia> number_of_parameters(g)
0

GenericCharacterTables.parameters — Function

parameters(t::CharTable)

Return all parameters the table t depends on.

Examples

julia> g=generic_character_table("GL2");

julia> parameters(g)
(q, (i, j, l, k))

parameters(t::SimpleCharTable)

Return all parameters the table t depends on.

Examples

julia> g=green_function_table("GL2");

julia> parameters(g)
(q, ())

parameters(class::GenericConjugacyClass)

Return the parameters of the conjugacy class type class. This includes the parameter names, ranges and exceptions.

Examples

julia> g=generic_character_table("GL2");

julia> parameters(conjugacy_class_type(g, 3))
i ∈ {1,…, q - 1}, j ∈ {1,…, q - 1} except i - j ∈ (q - 1)ℤ

parameters(class::SimpleGenericConjugacyClass)

Return the parameters of the conjugacy class type class. This includes the parameter names, ranges and exceptions.

Examples

julia> g=generic_character_table("uniGL2");

julia> parameters(conjugacy_class_type(g, 1))

parameters(t::Table, class::Int64)

Return the parameters of the class type class of the table t. This includes the parameter names, ranges and exceptions.

Examples

julia> g=generic_character_table("GL2");

julia> parameters(g, 3)
i ∈ {1,…, q - 1}, j ∈ {1,…, q - 1} except i - j ∈ (q - 1)ℤ

parameters(char::AbstractGenericCharacter)

Return the parameters of the character type char. This includes the parameter names, ranges and exceptions.

Examples

julia> g=generic_character_table("GL2");

julia> parameters(g[3])
k ∈ {1,…, q - 1}, l ∈ {1,…, q - 1} except -l + k ∈ (q - 1)ℤ

AbstractAlgebra.order — Method

order(t::Table)

Return the order of the table t.

Examples

julia> g=generic_character_table("GL2");

julia> order(g)
q^4 - q^3 - q^2 + q

Iteration

Tables implement Julia's iteration interface to iterate over the irreducible character types stored in the table. For a table T,

length(T) returns the number of character types in the table,

T[i] returns the $i$th character type.

For example we can use this to compute the order of the underlying group type. (Of course this can also be checked via order, which retrieves a precomputed value.)

julia> g=generic_character_table("GL2");

julia> sum(number_of_characters(c)*degree(c)^2 for c in g)
q^4 - q^3 - q^2 + q

julia> order(g)
q^4 - q^3 - q^2 + q

Tables also implement a part of Julia's interface for matrices. This can be used to extract "subtables" (although these are then of course merely matrices and not generic character table objects).

julia> g=generic_character_table("GL3");

julia> g[3:5,2:4]
3×3 Matrix{GenericCharacterTables.GenericCyclo}:
 0                               …  q*E(q - 1)^(2*a*n + b*n)
 (q + 1)*E(q - 1)^(a*m + 2*a*n)     E(q - 1)^(2*a*n + b*m) + (q + 1)*E(q - 1)^(a*m + a*n + b*n)
 q*E(q - 1)^(a*m + 2*a*n)           q*E(q - 1)^(2*a*n + b*m) + (q + 1)*E(q - 1)^(a*m + a*n + b*n)

Here we extracted the values of the character types 3, 4 and 5 on the conjugacy class types 2, 3 and 4. Using colons one can even get all values into a matrix.

julia> g=generic_character_table("GL2");

julia> g[:,:]
4×4 Matrix{GenericCharacterTables.GenericCyclo}:
 E(q - 1)^(2*i*k)              …  E(q - 1)^(i*k)
 q*E(q - 1)^(2*i*k)               -E(q - 1)^(i*k)
 (q + 1)*E(q - 1)^(i*l + i*k)     0
 (q - 1)*E(q - 1)^(i*k)           -E(q^2 - 1)^(q*i*k) - E(q^2 - 1)^(i*k)