Unexported Functions

These are all functions for internal use only:

GenericCharacterTables.normalize — Method

normalize(a::FracPoly)

Return a normal form b of a such that $\exp(2\pi \mathrm{i} \cdot a) = \exp(2\pi \mathrm{i} \cdot b)$.

This is done by reducing the numerators of the coefficients modulo their denominators.

GenericCharacterTables.shift_char_parameters — Method

shift_char_parameters(t::CharTable{T}, a::Union{Parameters{T},Cyclotomic{T}}, steps::Int64) where T <: NfPoly

Replace all character parameters of t in a by their counterparts suffixed with steps.

This is done by shifting them steps*nrparams(t) steps further in t.argumentring.

GenericCharacterTables.shift_class_parameters — Method

shift_class_parameters(t::CharTable{T}, a::Union{Parameters{T},Cyclotomic{T}}, steps::Int64) where T <: NfPoly

Replace all class parameters of t in a by their counterparts suffixed with steps.

This is done by shifting them steps*nrparams(t) steps further in t.argumentring.

GenericCharacterTables.shrink — Method

shrink(a::Set{ParameterException{<:NfPoly}})

Remove exceptions from a that follow from the others.

GenericCharacterTables.simplify — Method

simplify(x::Union{Cyclotomic{T}, ParameterException{T}}, t::CharTable{T}) where T <: NfPoly

Simplify x according to the congruence relation given by t.

GenericCharacterTables.simplify — Method

simplify(x::Union{Cyclotomic{T}, ParameterException{T}}, c::T, cinv::Generic.FracFieldElem{T}) where T <: NfPoly

Simplify x by replacing the generator of the polynomials of type T by c and then back to cinv.

This is used to remove terms that look fractional but are in fact whole.

GenericCharacterTables.try_convert_to_real — Method

try_convert_to_real(x::CycloFrac{<:NfPoly})

Try to convert x to a real number.

This only works if x is a fraction of trivial sums where all arguments are zero.

GenericCharacterTables.weak_sign — Method

weak_sign(a::QQFieldElem)

Return the sign of the leading coefficient of the polynomial representation of a.

GenericCharacterTables.weak_sign — Method

weak_sign(a::QQFieldElem)

Return the sign of a.