Unexported Functions

These are all functions for internal use only:

Base.iszero — Method

iszero(x::CycloFrac; ignore_exceptions::Bool=false)

Return if x is zero. If ignore_exceptions is true then the exceptions of x will not be considered.

Base.show — Method

show(io::IO, c::AbstractGenericCharacter)

Display a summary of the generic character c.

Examples

julia> g=genchartab("GL2");

julia> g[3]
Generic character of GL2
  with parameters
    k ∈ {1,…, q - 1}, l ∈ {1,…, q - 1} except l - k ∈ (q - 1)ℤ
  of degree q + 1
  with values
    (q + 1)*exp(2π𝑖(1//(q - 1)*i*l + 1//(q - 1)*i*k))
    exp(2π𝑖(1//(q - 1)*i*l + 1//(q - 1)*i*k))
    exp(2π𝑖(1//(q - 1)*i*k + 1//(q - 1)*j*l)) + exp(2π𝑖(1//(q - 1)*i*l + 1//(q - 1)*j*k))
    0

julia> [g[3]]
1-element Vector{GenericCharacterTables.GenericCharacter{QQPolyRingElem}}:
 Generic character of GL2

Base.show — Method

show(io::IO, t::Table)

Display a summary of the generic character table t.

Examples

julia> g=genchartab("GL2")
Generic character table GL2
  of order q^4 - q^3 - q^2 + q
  with 4 irreducible character types
  with 4 class types
  with parameters (i, j, l, k)

julia> [g]
1-element Vector{GenericCharacterTables.CharTable{QQPolyRingElem}}:
 Generic character table GL2

GenericCharacterTables.normal_form — Method

normal_form(a::FracPoly)

Return a normal form b of a such that $\exp(2\pi \mathrm{i} \cdot a) = \exp(2\pi \mathrm{i} \cdot b)$.

This is done by reducing the numerators of the coefficients modulo their denominators.

GenericCharacterTables.shift_char_parameters — Method

shift_char_parameters(t::CharTable{T}, a::Union{Parameters{T},Cyclotomic{T}}, steps::Int64) where T <: NfPoly

Replace all character parameters of t in a by their counterparts suffixed with steps.

This is done by shifting them steps*nrparams(t) steps further in t.argumentring.

GenericCharacterTables.shift_class_parameters — Method

shift_class_parameters(t::CharTable{T}, a::Union{Parameters{T},Cyclotomic{T}}, steps::Int64) where T <: NfPoly

Replace all class parameters of t in a by their counterparts suffixed with steps.

This is done by shifting them steps*nrparams(t) steps further in t.argumentring.

GenericCharacterTables.shrink — Method

shrink(a::CycloFrac{<:NfPoly})

Remove exceptions from a that follow from the others. And try to simplify the representation of a.

GenericCharacterTables.shrink — Method

shrink(a::Set{ParameterException{<:NfPoly}})

Remove exceptions from a that follow from the others.

GenericCharacterTables.simplify — Method

simplify(x::Union{Cyclotomic{T}, ParameterException{T}}, t::CharTable{T}) where T <: NfPoly

Simplify x according to the congruence relation given by t.

GenericCharacterTables.simplify — Method

simplify(x::Union{Cyclotomic{T}, ParameterException{T}}, c::T, cinv::Generic.FracFieldElem{T}) where T <: NfPoly

Simplify x by replacing the generator of the polynomials of type T by c and then back to cinv.

This is used to remove terms that look fractional but are in fact whole.

GenericCharacterTables.weak_sign — Method

weak_sign(a::QQFieldElem)

Return the sign of the leading coefficient of the polynomial representation of a.

GenericCharacterTables.weak_sign — Method

weak_sign(a::QQFieldElem)

Return the sign of a.